+++ app/(Base)/page.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 import TextArea from '@/components/TextArea';
 import UserSelectPallete from '@/components/UserSelectPallete';
 import ResetTestButton from '@/components/reset-test-button';
-import useTimer from '@/helpers/utils/useTimer';
+import useTimer from '@/hooks/useTimer';
 import { userConfigStore } from '@/store';
-import useKeydownGetter from '@/helpers/utils/useKeydownGetter';
+import useKeydownGetter from '@/hooks/useKeydownGetter';
 import { IoIosColorPalette } from 'react-icons/io';
 import ThemeChoose from '@/components/expo/theme-choose';
 
@@ -20,15 +20,6 @@ export default function HomePage() {
          <UserSelectPallete reset={reset} />
          <div className="flex items-center justify-between mt-16 sm:px-10">
             <div className="text-xl font-medium font-poppins text-accent flex-wrap">
-               {/* <button
-                  className={`flex font-bold items-center justify-center rounded py-2 mt-5 text-input w-48 ${
-                     isRunning || isExited ? 'bg-background' : 'bg-foreground'
-                  }`}
-                  onClick={run}
-                  disabled={isRunning || isExited}
-               >
-                  start
-               </button> */}
                <div className="mb-2">{timer}</div>
             </div>
