+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { Metadata } from 'next';
 import '@/styles/globals.css';
 import { Analytics } from '@vercel/analytics/react';
 import { Inconsolata } from 'next/font/google';
-import { ThemeProvider } from '@/helpers/lib/theme-provider';
+import { ThemeProvider } from '@/lib/theme-provider';
 import Head from 'next/head';
 import { THEMES } from '@/components/expo/theme.constant';
