+++ components/UI/select.tsx
@@ -19,14 +19,14 @@ const SelectTrigger = React.forwardRef<
    <SelectPrimitive.Trigger
       ref={ref}
       className={cn(
-         'flex h-9 w-full items-center justify-between whitespace-nowrap rounded-md bg-input/20 px-3 py-2 font-medium shadow-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring disabled:cursor-not-allowed disabled:opacity-50 [&>span]:line-clamp-1',
+         'flex h-9 w-full items-center justify-between border border-input whitespace-nowrap rounded-md bg-input/20 px-3 py-2 font-medium shadow-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring disabled:cursor-not-allowed disabled:opacity-50 [&>span]:line-clamp-1',
          className
       )}
       {...props}
    >
       {children}
       <SelectPrimitive.Icon asChild>
-         <ChevronDown className="h-5 w-5 opacity-50" />
+         <ChevronDown className="w-5 h-5 opacity-50" />
       </SelectPrimitive.Icon>
    </SelectPrimitive.Trigger>
 ));
@@ -125,7 +125,7 @@ const SelectItem = React.forwardRef<
    >
       <span className="absolute right-2 flex h-3.5 w-3.5 items-center justify-center">
          <SelectPrimitive.ItemIndicator>
-            <Check className="h-4 w-4" />
+            <Check className="w-4 h-4" />
          </SelectPrimitive.ItemIndicator>
       </span>
       <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
