+++ hooks/useTimer.ts
