+++ lib/reset.ts
